+++ src/main/java/ledger/database/storage/SQL/ISQLDatabasePayee.java
@@ -4,6 +4,8 @@ import ledger.database.entity.Payee;
 import ledger.database.entity.Tag;
 import ledger.database.storage.SQL.SQLite.ISQLiteDatabase;
 import ledger.database.storage.table.PayeeTable;
+import ledger.database.storage.table.TagTable;
+import ledger.database.storage.table.TagToPayeeTable;
 import ledger.exception.StorageException;
 
 import java.sql.PreparedStatement;
@@ -52,7 +54,7 @@ public interface ISQLDatabasePayee extends ISQLiteDatabase {
 
             if (payee.getTags() != null) {
                 for(Tag t : payee.getTags()) {
-                    insertTagIfNew(t);
+                    lookupAndInsertTag(t);
                     addTagForPayee(t, payee);
                 }
             }
@@ -93,7 +95,7 @@ public interface ISQLDatabasePayee extends ISQLiteDatabase {
             if (payee.getTags() != null) {
                 deleteAllTagsForPayee(payee);
                 for(Tag t : payee.getTags()) {
-                    insertTagIfNew(t);
+                    lookupAndInsertTag(t);
                     addTagForPayee(t, payee);
                 }
             }
@@ -123,8 +125,9 @@ public interface ISQLDatabasePayee extends ISQLiteDatabase {
                 int payeeID = rs.getInt(PayeeTable.PAYEE_ID);
                 String payeeName = rs.getString(PayeeTable.PAYEE_NAME);
                 String payeeDesc = rs.getString(PayeeTable.PAYEE_DESC);
+                List<Tag> tags = getAllTagsForPayeeId(payeeID);
 
-                Payee currentPayee = new Payee(payeeName, payeeDesc, payeeID);
+                Payee currentPayee = new Payee(payeeName, payeeDesc, payeeID, tags);
 
                 payeeList.add(currentPayee);
             }
@@ -134,7 +137,83 @@ public interface ISQLDatabasePayee extends ISQLiteDatabase {
         }
     }
 
-    default Tag insertTagIfNew(Tag currentTag) throws StorageException {
+    @Override
+    default List<Tag> getAllTagsForPayee(Payee payee) throws StorageException {
+        return getAllTagsForPayeeId(payee.getId());
+    }
+
+    @Override
+    default List<Tag> getAllTagsForPayeeId(int payeeId) throws StorageException{
+        try {
+            PreparedStatement stmt = getDatabase().prepareStatement("SELECT " + TagTable.TAG_ID +
+                    ", " + TagTable.TAG_NAME +
+                    ", " + TagTable.TAG_DESC +
+                    " FROM " + TagTable.TABLE_NAME + " WHERE " + TagTable.TAG_ID + " IN (SELECT " + TagToPayeeTable.TTPE_TAG_ID +
+
+                    " FROM " + TagToPayeeTable.TABLE_NAME +
+                    " WHERE " + TagToPayeeTable.TTPE_PAYEE_ID + " =?)");
+            stmt.setInt(1, payeeId);
+            ResultSet rs = stmt.executeQuery();
+            ArrayList<Tag> tags = new ArrayList<>();
+
+            while (rs.next()) {
+
+                int tagID = rs.getInt(TagTable.TAG_ID);
+                String tagName = rs.getString(TagTable.TAG_NAME);
+                String tagDescription = rs.getString(TagTable.TAG_DESC);
+
+                Tag currentTag = new Tag(tagName, tagDescription, tagID);
+
+                tags.add(currentTag);
+            }
+            return tags;
+        } catch (java.sql.SQLException e) {
+            throw new StorageException("Error while getting all tags for a payee", e);
+        }
+    }
+
+    @Override
+    default void deleteTagForPayee(Tag tag, Payee payee) throws StorageException {
+        payee = lookupAndInsertPayee(payee);
+        tag = lookupAndInsertTag(tag);
+        try {
+            PreparedStatement stmt = getDatabase().prepareStatement("DELETE FROM " + TagToPayeeTable.TABLE_NAME + " WHERE " + TagToPayeeTable.TTPE_PAYEE_ID + " =? AND " + TagToPayeeTable.TTPE_TAG_ID + "=?");
+            stmt.setInt(1, payee.getId());
+            stmt.setInt(2, tag.getId());
+            stmt.execute();
+        } catch (java.sql.SQLException e) {
+            throw new StorageException("Error while deleting a tag for a payee", e);
+        }
+    }
+
+    @Override
+    default void deleteAllTagsForPayee(Payee p) throws StorageException{
+        Payee payee = lookupAndInsertPayee(p);
+        try {
+            PreparedStatement stmt = getDatabase().prepareStatement("DELETE FROM " + TagToPayeeTable.TABLE_NAME + " WHERE " + TagToPayeeTable.TTPE_PAYEE_ID + " =?");
+            stmt.setInt(1, payee.getId());
+            stmt.execute();
+        } catch (java.sql.SQLException e) {
+            if (payee == null)  throw new StorageException("Error while deleting all tags for payee", e);
+            else throw new StorageException("Error while deleting all tags for Payee " + payee.getName(), e);
+        }
+    }
+
+    @Override
+    default void addTagForPayee(Tag tag, Payee payee) throws StorageException {
+        payee = lookupAndInsertPayee(payee);
+        tag = lookupAndInsertTag(tag);
+        try {
+            PreparedStatement stmt = getDatabase().prepareStatement("INSERT INTO " + TagToPayeeTable.TABLE_NAME + " ( " + TagToPayeeTable.TTPE_TAG_ID + ", " + TagToPayeeTable.TTPE_PAYEE_ID + " ) VALUES (?,?)");
+            stmt.setInt(1, tag.getId());
+            stmt.setInt(2, payee.getId());
+            stmt.execute();
+        } catch (java.sql.SQLException e) {
+            throw new StorageException("Error while adding a tag for a payee", e);
+        }
+    }
+
+    default Tag lookupAndInsertTag(Tag currentTag) throws StorageException {
         Tag existingTag;
         if (currentTag.getId() != -1) {
             existingTag = currentTag;
@@ -145,8 +224,25 @@ public interface ISQLDatabasePayee extends ISQLiteDatabase {
             currentTag.setId(existingTag.getId());
         } else {
             insertTag(currentTag);
+            currentTag = getTagForNameAndDescription(currentTag.getName(), currentTag.getDescription());
         }
         return currentTag;
+    }
+
+    default Payee lookupAndInsertPayee(Payee currentPayee) throws StorageException {
+        Payee existingPayee;
+        if (currentPayee.getId() != -1) {
+            existingPayee = currentPayee;
+        } else {
+            existingPayee = getPayeeForNameAndDescription(currentPayee.getName(), currentPayee.getDescription());
+        }
+        if (existingPayee != null) {
+            currentPayee.setId(existingPayee.getId());
+        } else {
+            insertPayee(currentPayee);
+            currentPayee = getPayeeForNameAndDescription(currentPayee.getName(), currentPayee.getDescription());
+        }
+        return currentPayee;
     }
 
 }
