+++ src/main/java/ledger/database/storage/SQL/ISQLDatabaseTag.java
@@ -110,107 +110,4 @@ public interface ISQLDatabaseTag extends ISQLiteDatabase {
             throw new StorageException("Error while getting all tags", e);
         }
     }
-
-    @Override
-    default List<Tag> getAllTagsForPayee(Payee payee) throws StorageException {
-        try {
-            PreparedStatement stmt = getDatabase().prepareStatement("SELECT " + TagTable.TAG_ID +
-                    ", " + TagTable.TAG_NAME +
-                    ", " + TagTable.TAG_DESC +
-                    " FROM " + TagTable.TABLE_NAME + " WHERE " + TagTable.TAG_ID + " IN (SELECT " + TagToPayeeTable.TTPE_TAG_ID +
-
-                    " FROM " + TagToPayeeTable.TABLE_NAME +
-                    " WHERE " + TagToPayeeTable.TTPE_PAYEE_ID + " =?)");
-            stmt.setInt(1, lookupAndInsertPayee(payee).getId());
-            ResultSet rs = stmt.executeQuery();
-            ArrayList<Tag> tags = new ArrayList<>();
-
-            while (rs.next()) {
-
-                int tagID = rs.getInt(TagTable.TAG_ID);
-                String tagName = rs.getString(TagTable.TAG_NAME);
-                String tagDescription = rs.getString(TagTable.TAG_DESC);
-
-                Tag currentTag = new Tag(tagName, tagDescription, tagID);
-
-                tags.add(currentTag);
-            }
-            return tags;
-        } catch (java.sql.SQLException e) {
-            throw new StorageException("Error while getting all tags for a payee", e);
-        }
-    }
-
-    @Override
-    default void deleteTagForPayee(Tag tag, Payee payee) throws StorageException {
-        payee = lookupAndInsertPayee(payee);
-        tag = lookupAndInsertTag(tag);
-        try {
-            PreparedStatement stmt = getDatabase().prepareStatement("DELETE FROM " + TagToPayeeTable.TABLE_NAME + " WHERE " + TagToPayeeTable.TTPE_PAYEE_ID + " =? AND " + TagToPayeeTable.TTPE_TAG_ID + "=?");
-            stmt.setInt(1, payee.getId());
-            stmt.setInt(2, tag.getId());
-            stmt.execute();
-        } catch (java.sql.SQLException e) {
-            throw new StorageException("Error while deleting a tag for a payee", e);
-        }
-    }
-
-    @Override
-    default void deleteAllTagsForPayee(Payee p) throws StorageException{
-        Payee payee = lookupAndInsertPayee(p);
-        try {
-            PreparedStatement stmt = getDatabase().prepareStatement("DELETE FROM " + TagToPayeeTable.TABLE_NAME + " WHERE " + TagToPayeeTable.TTPE_PAYEE_ID + " =?");
-            stmt.setInt(1, payee.getId());
-            stmt.execute();
-        } catch (java.sql.SQLException e) {
-            if (payee == null)  throw new StorageException("Error while deleting all tags for payee", e);
-            else throw new StorageException("Error while deleting all tags for Payee " + payee.getName(), e);
-        }
-    }
-
-    @Override
-    default void addTagForPayee(Tag tag, Payee payee) throws StorageException {
-        payee = lookupAndInsertPayee(payee);
-        tag = lookupAndInsertTag(tag);
-        try {
-            PreparedStatement stmt = getDatabase().prepareStatement("INSERT INTO " + TagToPayeeTable.TABLE_NAME + " ( " + TagToPayeeTable.TTPE_TAG_ID + ", " + TagToPayeeTable.TTPE_PAYEE_ID + " ) VALUES (?,?)");
-            stmt.setInt(1, tag.getId());
-            stmt.setInt(2, payee.getId());
-            stmt.execute();
-        } catch (java.sql.SQLException e) {
-            throw new StorageException("Error while adding a tag for a payee", e);
-        }
-    }
-
-    default Tag lookupAndInsertTag(Tag currentTag) throws StorageException {
-        Tag existingTag;
-        if (currentTag.getId() != -1) {
-            existingTag = currentTag;
-        } else {
-            existingTag = getTagForNameAndDescription(currentTag.getName(), currentTag.getDescription());
-        }
-        if (existingTag != null) {
-            currentTag.setId(existingTag.getId());
-        } else {
-            insertTag(currentTag);
-            currentTag = getTagForNameAndDescription(currentTag.getName(), currentTag.getDescription());
-        }
-        return currentTag;
-    }
-
-    default Payee lookupAndInsertPayee(Payee currentPayee) throws StorageException {
-        Payee existingPayee;
-        if (currentPayee.getId() != -1) {
-            existingPayee = currentPayee;
-        } else {
-            existingPayee = getPayeeForNameAndDescription(currentPayee.getName(), currentPayee.getDescription());
-        }
-        if (existingPayee != null) {
-            currentPayee.setId(existingPayee.getId());
-        } else {
-            insertPayee(currentPayee);
-            currentPayee = getPayeeForNameAndDescription(currentPayee.getName(), currentPayee.getDescription());
-        }
-        return currentPayee;
-    }
 }
